inventory features added: color change when low inventory overall better layout

+++ demo/src/main/java/projectone/demo/model/Products.java
@@ -20,7 +20,7 @@ import jakarta.persistence.Table;
 public class Products implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger product_id;
+    private Long product_id;
 
     private String productname;
